moved internal condition logic to its own package
While implementing the Parameterize
functions for SimpleCondition an import
cycle was encountered when importing
the column parser. So, these changes were made
