+++ builders/builders.go
@@ -1,8 +1,6 @@
 package builders
 
-import (
-	intypes "github.com/williabk198/jagsqlb/internal/types"
-)
+import inconds "github.com/williabk198/jagsqlb/internal/conditions"
 
 type Builder interface {
 	Build() (query string, queryParams []any, err error)
@@ -11,11 +9,11 @@ type Builder interface {
 type SelectBuilder interface {
 	Builder
 	Table(table string, columns ...string) SelectBuilder
-	Where(intypes.Condition, ...intypes.Condition) WhereBuilder
+	Where(inconds.Condition, ...inconds.Condition) WhereBuilder
 }
 
 type WhereBuilder interface {
 	Builder
-	And(intypes.Condition, ...intypes.Condition) WhereBuilder
-	Or(intypes.Condition, ...intypes.Condition) WhereBuilder
+	And(inconds.Condition, ...inconds.Condition) WhereBuilder
+	Or(inconds.Condition, ...inconds.Condition) WhereBuilder
 }
