+++ conditions/conditions_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	intypes "github.com/williabk198/jagsqlb/internal/types"
+	inconds "github.com/williabk198/jagsqlb/internal/conditions"
 )
 
 func TestEquals(t *testing.T) {
@@ -15,7 +15,7 @@ func TestEquals(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success",
@@ -23,7 +23,7 @@ func TestEquals(t *testing.T) {
 				columnName: "col1",
 				value:      "testing",
 			},
-			want: intypes.SimpleCondition{
+			want: inconds.SimpleCondition{
 				ColumnName: "col1",
 				Operator:   "=",
 				Values:     []any{"testing"},
@@ -45,7 +45,7 @@ func TestNotEquals(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success",
@@ -53,7 +53,7 @@ func TestNotEquals(t *testing.T) {
 				columnName: "col1",
 				value:      "testing",
 			},
-			want: intypes.SimpleCondition{
+			want: inconds.SimpleCondition{
 				ColumnName: "col1",
 				Operator:   "!=",
 				Values:     []any{"testing"},
@@ -75,7 +75,7 @@ func TestGraterThan(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success",
@@ -83,7 +83,7 @@ func TestGraterThan(t *testing.T) {
 				columnName: "col1",
 				value:      42,
 			},
-			want: intypes.SimpleCondition{
+			want: inconds.SimpleCondition{
 				ColumnName: "col1",
 				Operator:   ">",
 				Values:     []any{42},
@@ -105,7 +105,7 @@ func TestGreaterThanEqual(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success",
@@ -113,7 +113,7 @@ func TestGreaterThanEqual(t *testing.T) {
 				columnName: "col1",
 				value:      42,
 			},
-			want: intypes.SimpleCondition{
+			want: inconds.SimpleCondition{
 				ColumnName: "col1",
 				Operator:   ">=",
 				Values:     []any{42},
@@ -135,7 +135,7 @@ func TestLessThan(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success",
@@ -143,7 +143,7 @@ func TestLessThan(t *testing.T) {
 				columnName: "col1",
 				value:      42,
 			},
-			want: intypes.SimpleCondition{
+			want: inconds.SimpleCondition{
 				ColumnName: "col1",
 				Operator:   "<",
 				Values:     []any{42},
@@ -165,7 +165,7 @@ func TestLessThanEqual(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success",
@@ -173,7 +173,7 @@ func TestLessThanEqual(t *testing.T) {
 				columnName: "col1",
 				value:      42,
 			},
-			want: intypes.SimpleCondition{
+			want: inconds.SimpleCondition{
 				ColumnName: "col1",
 				Operator:   "<=",
 				Values:     []any{42},
@@ -195,7 +195,7 @@ func TestIn(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success",
@@ -203,7 +203,7 @@ func TestIn(t *testing.T) {
 				columnName: "col1",
 				value:      []any{42, 56, 127},
 			},
-			want: intypes.SimpleCondition{
+			want: inconds.SimpleCondition{
 				ColumnName: "col1",
 				Operator:   "IN",
 				Values:     []any{42, 56, 127},
@@ -225,7 +225,7 @@ func TestNotIn(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success",
@@ -233,7 +233,7 @@ func TestNotIn(t *testing.T) {
 				columnName: "col1",
 				value:      []any{42, 56, 127},
 			},
-			want: intypes.SimpleCondition{
+			want: inconds.SimpleCondition{
 				ColumnName: "col1",
 				Operator:   "NOT IN",
 				Values:     []any{42, 56, 127},
@@ -256,7 +256,7 @@ func TestBetween(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success",
@@ -265,7 +265,7 @@ func TestBetween(t *testing.T) {
 				val1:       42,
 				val2:       56,
 			},
-			want: intypes.SimpleCondition{
+			want: inconds.SimpleCondition{
 				ColumnName: "col1",
 				Operator:   "BETWEEN",
 				Values:     []any{42, 56},
@@ -288,7 +288,7 @@ func TestNotBetween(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success",
@@ -297,7 +297,7 @@ func TestNotBetween(t *testing.T) {
 				val1:       42,
 				val2:       56,
 			},
-			want: intypes.SimpleCondition{
+			want: inconds.SimpleCondition{
 				ColumnName: "col1",
 				Operator:   "NOT BETWEEN",
 				Values:     []any{42, 56},
@@ -313,9 +313,9 @@ func TestNotBetween(t *testing.T) {
 
 func TestGroupedAnd(t *testing.T) {
 	type args struct {
-		cond1           intypes.Condition
-		cond2           intypes.Condition
-		additionalConds []intypes.Condition
+		cond1           inconds.Condition
+		cond2           inconds.Condition
+		additionalConds []inconds.Condition
 	}
 
 	testCond1 := LessThanEqual("t1.col1", 42)
@@ -325,7 +325,7 @@ func TestGroupedAnd(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success; Minimal",
@@ -333,9 +333,9 @@ func TestGroupedAnd(t *testing.T) {
 				cond1: testCond1,
 				cond2: testCond2,
 			},
-			want: intypes.GroupedConditions{
+			want: inconds.GroupedConditions{
 				Conjunction: " AND ",
-				Conditions:  []intypes.Condition{testCond1, testCond2},
+				Conditions:  []inconds.Condition{testCond1, testCond2},
 			},
 		},
 		{
@@ -343,11 +343,11 @@ func TestGroupedAnd(t *testing.T) {
 			args: args{
 				cond1:           testCond1,
 				cond2:           testCond2,
-				additionalConds: []intypes.Condition{testCond3},
+				additionalConds: []inconds.Condition{testCond3},
 			},
-			want: intypes.GroupedConditions{
+			want: inconds.GroupedConditions{
 				Conjunction: " AND ",
-				Conditions:  []intypes.Condition{testCond1, testCond2, testCond3},
+				Conditions:  []inconds.Condition{testCond1, testCond2, testCond3},
 			},
 		},
 	}
@@ -360,9 +360,9 @@ func TestGroupedAnd(t *testing.T) {
 
 func TestGroupedOr(t *testing.T) {
 	type args struct {
-		cond1           intypes.Condition
-		cond2           intypes.Condition
-		additionalConds []intypes.Condition
+		cond1           inconds.Condition
+		cond2           inconds.Condition
+		additionalConds []inconds.Condition
 	}
 
 	testCond1 := LessThanEqual("t1.col1", 42)
@@ -372,7 +372,7 @@ func TestGroupedOr(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want intypes.Condition
+		want inconds.Condition
 	}{
 		{
 			name: "Success; Minimal",
@@ -380,9 +380,9 @@ func TestGroupedOr(t *testing.T) {
 				cond1: testCond1,
 				cond2: testCond2,
 			},
-			want: intypes.GroupedConditions{
+			want: inconds.GroupedConditions{
 				Conjunction: " OR ",
-				Conditions:  []intypes.Condition{testCond1, testCond2},
+				Conditions:  []inconds.Condition{testCond1, testCond2},
 			},
 		},
 		{
@@ -390,11 +390,11 @@ func TestGroupedOr(t *testing.T) {
 			args: args{
 				cond1:           testCond1,
 				cond2:           testCond2,
-				additionalConds: []intypes.Condition{testCond3},
+				additionalConds: []inconds.Condition{testCond3},
 			},
-			want: intypes.GroupedConditions{
+			want: inconds.GroupedConditions{
 				Conjunction: " OR ",
-				Conditions:  []intypes.Condition{testCond1, testCond2, testCond3},
+				Conditions:  []inconds.Condition{testCond1, testCond2, testCond3},
 			},
 		},
 	}
