+++ internal/conditions/grouped.go
@@ -1,20 +1,4 @@
-package intypes
-
-// Condition is essentially the same as a Builder, but needed a way to differentiate a Condtion from a Builder
-type Condition interface {
-	Parameterize() (string, []any, error)
-}
-
-type SimpleCondition struct {
-	ColumnName string
-	Operator   string
-	Values     []any
-}
-
-func (sc SimpleCondition) Parameterize() (string, []any, error) {
-	// TODO: The returned string should be of the format "<column> <operator> ?"
-	panic("unimplemented")
-}
+package inconds
 
 type GroupedConditions struct {
 	Conjunction string // This value should always be either " AND " or " OR "
